feat: 增加菜单树接口

+++ internal/app/model/base.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Base struct {
-	ID        uint           `gorm:"primarykey" json:"id"`
+	ID        uint64         `gorm:"primarykey" json:"id"`
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `gorm:"index" json:"-"`
